% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupBoost-package.R
\docType{package}
\name{SubgroupBoost-package}
\alias{SubgroupBoost}
\alias{SubgroupBoost-package}
\title{SubgroupBoost: Value function Guided Subgrouping via Gradient Tree Boosting}
\description{
This package is to identify the treatment benefiting and non-benefiting subgroups based on baseline characteristics. It searches for subgorup membership scores by maximizing a RMST or win-ratio based value function that directly reflects the subgroup-treatment interaction effect via gradient tree boosting.
}
\author{
\strong{Maintainer}: Peng Liu \email{peng.liu9@merck.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
