% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupBoost.RMST.R
\name{SubgroupBoost.RMST}
\alias{SubgroupBoost.RMST}
\title{Title RMST based value function guided subgrouping}
\usage{
SubgroupBoost.RMST(dat)
}
\arguments{
\item{dat}{a dataframe, rows are subject and columns are predictors, treatment indicator variable \code{trt01p},
event indicator variable \code{evnt}, and time variable \code{aval}. \code{trt01p} is coded as -1 (control) or 1 (treatment),and \code{evnt} is coded as 0
(censored) or 1 (event).}
}
\value{
a fitted gradient boosting model
}
\description{
Title RMST based value function guided subgrouping
}
\details{
add details later
}
\examples{
NULL
}
