% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupBoost.wd.R
\name{SubgroupBoost.wd}
\alias{SubgroupBoost.wd}
\title{Title Win-difference based value function guided subgrouping}
\usage{
SubgroupBoost.wd(dat, info, comparison)
}
\arguments{
\item{dat}{a dataframe, rows are subject and columns are predictors,
(censored) or 1 (event).}

\item{info}{a dataframe of treatment indicator variable \code{trt01p}, event indicator variable \code{evnt},
and time variable \code{aval}. \code{trt01p} is coded as -1 (control) or 1 (treatment),and \code{evnt} is coded as 0
(censored) or 1(event). If there are two survival outcomes, use \code{evnt1},\code{evnt2} and \code{aval1},\code{aval2}
to distinguish the two outcomes.}

\item{comparison}{The types of outcome to compare in win-difference based value function. The possible choices are
\itemize{
\item{\code{survival survival}}{Two time-to-event outcomes}
\item{\code{survival}}{one time-to-event outcome}
\item{\code{continous}}{one continous outcome, larger value means more benefit from treatment}
\item{\code{binary}}{one binary outcome, larger value means more benefit from treatment}
\item{\code{ordinal}}{one ordinal discrete outcome, larger value means more benefit from treatment}
}}
}
\value{
a fitted gradient boosting model
}
\description{
Title Win-difference based value function guided subgrouping
}
\details{
add details later
}
\examples{
NULL
}
